LPS-84675 Move to display package as in com.liferay.configuration.admin.web.internal.display
