+++ modules/apps/sharing/sharing-web/src/main/java/com/liferay/sharing/web/internal/portlet/action/ShareEntryMVCActionCommand.java
@@ -25,7 +25,7 @@ import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.sharing.service.SharingEntryService;
-import com.liferay.sharing.web.internal.constants.SharingEntryPermissionDisplayActionKey;
+import com.liferay.sharing.web.internal.display.SharingEntryPermissionDisplayActionKey;
 import com.liferay.sharing.web.internal.constants.SharingPortletKeys;
 
 import javax.portlet.ActionRequest;
